модули не работают
доработка модулей (программа не работает)
добавлены часть шаблонов и дефолтных выводов
