+++ default/materials_default
@@ -1,0 +1,33 @@
+% --- Zr-Nb cladding and shroud tube:
+
+mat clad    -6.55000 rgb 100 100 100
+40000.06c   -0.99000
+41093.06c   -0.01000
+
+mat tube    -6.58000 rgb 200 200 200
+40000.06c   -0.97500
+41093.06c   -0.02500
+
+mat wst -4.100 rgb 150 150 255 %12X18H10T
+6012.06c 0.00600
+26056.06c 0.34900
+24052.06c 0.09000
+28059.06c 0.05000
+22048.06c 0.00500
+8016.06c 0.44444
+1001.06c 0.05556
+
+
+% --- Water:
+
+mat water   -0.7207  moder lwtr 1001 rgb 0 0 255
+ 1001.06c    2.0\n 8016.06c    1.0
+ 
+
+ % --- UO2 fuel enriched to 3.6 wt-% U-235 with gd-64:
+ 
+ mat fuel_gd   -10.45700 rgb 200 255 200
+ 92235.09c   -0.04103
+ 92238.09c   -0.79631
+ 64000.09c   -0.05000
+ 8016.09c   -0.11266
