+++ examples/template
@@ -1,0 +1,1 @@
+nine of {i} dantists recomend
