--- modules/output/materials/materials_default.py
@@ -1,3 +1,0 @@
-
-def materials_standart():
-    return 0
